Add lexing and parsing.
It is still far from perfect. The .y file has been written based on
Annex A of the standard, so it should contain most syntactical elements.
Currently we only construct the AST from a single compound statement.
We have also implemented a debug print function for the AST, that should
output (mostly) valid C code.